refactor: @Auth 위치 변경
